box: fix replica hang after applying already set name
When the name is manually set on master by replace in _cluster space,
calling box.cfg on replica with the same name causes its hang. The
problem is the fact, that resubscribe is initiated and waiting for
APPLIER_REGISTERED status is started. As applier knows, that no
registration should be done, this never happens.

Let's don't initiate registration, when instance name is already set.

Needed for #8978

NO_DOC=bugfix
NO_CHANGELOG=not released yet