lua: move mkversion into a separate module
This commit moves all code, related to working with versions and which
was used in box/lua/upgrade.lua, into a separate module and exports it
to Lua API as 'internal.version'

This is needed, as in the following commit we set names automatically
only when schema version is more than 3.0.0. This module is used their
in order to avoid code duplication.

Follow-up #8978

NO_DOC=internal
NO_TEST=<already tested>
NO_CHANGELOG=internal