relay: check return value of xrow_decode_synchro
Relay sometimes decodes the PROMOTE packets to be sent in order
to conditionally delay their dispatch. It was believed that own WALs
can't be corrupted and hence there is no point in checking that decoding
succeeds.

Let's be more strict here and check the decoding result before
proceeding.

Closes #9265

NO_DOC=bugfix
NO_TEST=hard to test
NO_CHANGELOG=not user-visible