test: add a fuzz-test for the transactional DDL
The test performs a random set of DDL operations in transactions. The
idea is to catch possible crashes on specific sets of operations.

The test is made as reproducible as possible by using the custom RNG
and making lua table traversals deterministic (by sorting them). So
the random seed is universal accross luajit versions and operating
systems.

The DDL operation log is a valid lua code that can be used to create
a reproduce manually, except the lines are executed in different
fibers, so it sometimes may be not sufficient to just copy them. The
preferred way to reproduce an issue is specifying seed though.

Closes #4349

NO_DOC=test
NO_CHANGELOG=test