journal: broadcast box.wal_error event on write error
Closes #9405

@TarantoolBot document
Title: Document then new built-in system event `box.wal_error`

The new event is broadcast whenever Tarantool fails to commit
a transaction to the write-ahead log (WAL), which usually means
there's a problem with the underlying disk storage. The new event's
payload is a table that currently contains the only field `count` that
stores the number of WAL errors happened so far or nil if there hasn't
been any WAL errors.