test/config: add simple configuration builder
It allows to construct a declarative configuration for a test case using
less boilerplace code/options, especially when a replicaset is to be
tested, not a single instance. See a description at top of the file for
details.

NO_DOC=testing helper
NO_CHANGELOG=see NO_DOC
NO_TEST=see NO_DOC