config: support master auto-discovery in vshard
This patch adds support for automatic master discovery in vshard. There
is no longer a need to reapply the vshard storage configuration every
time an instance becomes a master or ceases to be a master.

Automatic master discovery also solves the problem with the rebalancer.
Previously, the rebalancer could not work correctly if the masters of
some replicasets were unknown, and since the vshard config generated by
the config module did not contain information about all the masters, the
rebalancer was disabled. The rebalancer can now perform master discovery
on its own, which is why it is enabled.

Before this patch, the config module was based on the
"test: use dofile() for configs instead of require" commit in vshard.
At a minimum, commit "storage: fix assertion error in conn_manager" is
now required.

Part of #8862

NO_DOC=will be added along with documentation for the rebalancer role
NO_CHANGELOG=will be added along with changelog for the rebalancer role