uri: add login and password fields
Closes #9435

@TarantoolBot document
Title: `login` and `password` fields in URI table

Now `uri.parse()` can parse a table containing the `login` and
`password` fields. Values from these fields take precedence over values
obtained from the string URI. For example, login and password of
`{uri = 'one:two:localhost:3301, login = 'alpha', password = 'omega'}`
will be `alpha` and `omega` respectively. If the `login` field is set
and the `password` field is not set, the password is set to `nil`.
If the `password` field is set, the `login` field must be present.