config: remove config.version option
The idea of `config.version` is to differentiate a configuration that is
written for the given tarantool version (and any unknown field is to be
considered as an error) from a configuration that deliberately contains
options that are unknown on the given tarantool version (to apply them
on newer tarantool versions).

This option was planned as mandatory one: a user must set it in its
config. OTOH, a user likely copy-pastes some baseline configuration and
adds some options from documentation pages or another project. The user
unlikely really want to pay attention to a supported version range for
each of used options to deduce a right `config.version` value.

The problem is mostly about conflict of a strict validation (no unknown
options) with schema evolution (ignore options for newer tarantool
versions). See #9452 for possible approaches to solve it.

One of the proposed mechaniscms to support schema evolution together
with a strict validation is implemented in next commits of the series.

Part of #9452

@TarantoolBot document
Title: config: `config.version` is removed

Another schema evolution mechanism is implemented in
https://github.com/tarantool/tarantool/issues/9452.