config: add version expression parsing
The usual expression is a kind of `tarantool_version >= 3.1.0 &&
tarantool_version < 4.0.0`, but all the comparison operators are
supported, logical and/or and grouping. See the previous commit for the
list of operators.

The module implements Pratt's expression parsing algorithm.

Part of #9452

NO_DOC=It is a supplementary module for config's conditional section
       predicates. To be documented in the last commit of the series.
NO_CHANGELOG=see NO_DOC