lua: fix loader behaviour in case of syntax error
The patch fixes error handling between module "searching" and "loading"
processes. Originally, if a module is not found, the error message is
returned to <require> and is appended to the "not found" message. If no
suitable module is found, the error is raised by <require>. In case,
when the desired module is found, <require> tries to load it. If any
error occurs while loading the module, the "loader" yields the
corresponding error. The latter part was broken since the commit
5be3a82be52c48591f1da9254b0f37fc2bb584a5 ("lua: fix loaders behavior"),
so if the error occurs while loading the module, Tarantool proceeds with
searching the module in other locations.

This bug was found via the lua-Harness suite[1] used in the
tarantool/luajit testing routine, so the test added in the patch
partially duplicates the particular test case in lua-Harness.

[1]: https://fperrad.frama.io/lua-Harness/

Needed for tarantool/tarantool-ee#585

NO_DOC=bugfix

(cherry picked from commit b99a35412d0c65dd84209872654fec9aaf2f1d02)